rename ws mod to dispatch
