+++ server/src/dispatch.rs
@@ -2,16 +2,64 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use std::collections::HashMap;
+use std::{collections::HashMap, sync::Arc};
 
-use phnxbackend::qs::{Notification, NotifierError};
+use phnxbackend::qs::{Notification, Notifier, NotifierError, grpc::GrpcListen};
 use phnxprotos::{
     convert::RefInto,
     queue_service::v1::{QueueEvent, QueueEventPayload, QueueEventUpdate, queue_event},
 };
 use phnxtypes::{identifiers::QsClientId, messages::client_ds::DsEventMessage};
-use tokio::sync::mpsc;
+use tokio::{
+    self,
+    sync::{Mutex, mpsc},
+};
 use tracing::info;
+
+/// This is a wrapper for dispatch actor that can be used to send out a
+/// notification over the dispatch.
+#[derive(Clone, Debug, Default)]
+pub struct DispatchNotifier {
+    pub dispatch: Arc<Mutex<Dispatch>>,
+}
+
+impl DispatchNotifier {
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
+
+impl Notifier for DispatchNotifier {
+    /// Notify a client that opened a websocket connection to the QS.
+    ///
+    /// # Arguments
+    /// queue_id - The queue ID of the client
+    /// ws_notification - The notification to send
+    ///
+    /// # Returns
+    ///
+    /// Returns `()` of the operation was successful and
+    /// `WebsocketNotifierError::ClientNotFound` if the client was not found.
+    async fn notify(
+        &self,
+        queue_id: &QsClientId,
+        notification: Notification,
+    ) -> Result<(), NotifierError> {
+        let mut dispatch = self.dispatch.lock().await;
+        dispatch.notify_client(queue_id, notification)
+    }
+}
+
+impl GrpcListen for DispatchNotifier {
+    async fn register_connection(
+        &self,
+        queue_id: QsClientId,
+        tx: mpsc::UnboundedSender<QueueEvent>,
+    ) {
+        let mut dispatch = self.dispatch.lock().await;
+        dispatch.connect(queue_id, tx);
+    }
+}
 
 /// Dispatch for all connections.
 ///
