+++ server/src/main.rs
@@ -8,12 +8,12 @@ use phnxbackend::{auth_service::AuthService, ds::Ds, infra_service::InfraService
 use phnxserver::{
     RateLimitsConfig, ServerRunParams,
     configurations::*,
+    dispatch::DispatchNotifier,
     enqueue_provider::SimpleEnqueueProvider,
     network_provider::MockNetworkProvider,
     push_notification_provider::ProductionPushNotificationProvider,
     run,
     telemetry::{get_subscriber, init_subscriber},
-    ws::DispatchNotifier,
 };
 use phnxtypes::identifiers::Fqdn;
 use tracing::info;
